Modeling tools.
The new Modeling module gives a new TH command `modelsFor`, which takes
a [Dec] and generates model code from it.

For each typesig `d :: T -> X`, we generate `model_d :: ModelOf T ->
ModelOf X`. In the definitions, we then remove any calls to `model`
and `unmodel`, and replace any other function calls in the same decs
group to their modeled version. This way a well-typed expression in the
dec stays well-typed in the result, but we've inlined all of the
modeling code. This is necessary because `unmodel` can't be implemented
in general.

Additionally, we simply re-emit the decs themselves (with model and
unmodel present) under new names, to ensure they typecheck beforehand.