Generate model tests
Now that we can create models, we should be able to run our tests over
the model. This change does that, by generating two sets of tests ---
one for the true functions, and one for the modeled ones.

Right now there is a good deal of jankiness to get everything to place
nicely. We manipulate the global Q state to pass data from one splice to
another --- namely the test decs generated via `theoremsOf`, and the
modeled name from `modelsFor`.