Replace ModelOf types with their actual models
th-tc has a hard time with type families and type synonyms. So rather
than replacing a type T in a model with `ModelOf T`, we instead replace
it with `μ T`. This means we're the ones doing the hard work and the
type families never appear in the generated code.

Everything works out though, because the types we're spitting out are
equivalent to the modelofs; so GHC can unify them.